refactor: renaming folders & reordering files
