+++ app/src/main/java/com/arkan/terbangin/presentation/auth/register/RegisterActivity.kt
@@ -1,23 +1,21 @@
-package com.arkan.terbangin.presentation.login
+package com.arkan.terbangin.presentation.auth.register
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.arkan.terbangin.R
-import com.arkan.terbangin.databinding.ActivityLoginBinding
-import com.arkan.terbangin.presentation.register.RegisterActivity
+import com.arkan.terbangin.databinding.ActivityRegisterBinding
+import com.arkan.terbangin.presentation.auth.login.LoginActivity
 
-class LoginActivity : AppCompatActivity() {
-    private val binding: ActivityLoginBinding by lazy {
-        ActivityLoginBinding.inflate(layoutInflater)
+class RegisterActivity : AppCompatActivity() {
+    private val binding: ActivityRegisterBinding by lazy {
+        ActivityRegisterBinding.inflate(layoutInflater)
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
         setContentView(binding.root)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
@@ -28,14 +26,14 @@ class LoginActivity : AppCompatActivity() {
     }
 
     private fun setClickListener() {
-        binding.tvNavToRegister.setOnClickListener {
-            navigateToRegister()
+        binding.tvNavToLogin.setOnClickListener {
+            navigateToLogin()
         }
     }
 
-    private fun navigateToRegister() {
+    private fun navigateToLogin() {
         startActivity(
-            Intent(this, RegisterActivity::class.java).apply {
+            Intent(this, LoginActivity::class.java).apply {
                 flags = Intent.FLAG_ACTIVITY_SINGLE_TOP or Intent.FLAG_ACTIVITY_CLEAR_TOP
             },
         )
