+++ app/src/main/java/com/arkan/terbangin/presentation/auth/reset_password/ResetPasswordActivity.kt
@@ -1,7 +1,6 @@
-package com.arkan.terbangin.presentation.resetpassword
+package com.arkan.terbangin.presentation.auth.reset_password
 
 import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
@@ -10,7 +9,6 @@ import com.arkan.terbangin.R
 class ResetPasswordActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
         setContentView(R.layout.activity_reset_password)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
