+++ app/src/main/res/layout/activity_flight_search.xml
@@ -5,7 +5,7 @@
     android:id="@+id/flight_search"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.flightsearch.FlightSearchActivity">
+    tools:context=".presentation.flight_search.FlightSearchActivity">
 
     <include
         android:id="@+id/layout_app_bar"
