+++ app/src/main/res/layout/activity_reset_password.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.resetpassword.ResetPasswordActivity">
+    tools:context=".presentation.auth.reset_password.ResetPasswordActivity">
 
     <ImageButton
         android:id="@+id/ib_btn_back"
