+++ app/src/main/res/layout/bottom_sheet_terminal_search.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentation.terminalsearch.TerminalSearchBottomSheet">
+    tools:context=".presentation.terminal_search.TerminalSearchBottomSheet">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
